changed long to int in giId

+++ src/main/java/com/goit/hotelonlinebooking/dao/ById.java
@@ -2,5 +2,5 @@ package com.goit.hotelonlinebooking.dao;
 
 public abstract class ById {
 
-   public abstract long getId();
+   public abstract int getId();
 }
